Added 2 floats for maximum probe area for cartesians and setting the max/min in an intelligent manner. Removed some debugging code :)
added do_home to make homing at the beginning of findBed() optional, and is_square to turn on probing corners.
x_max and y_max are new config options for this levelling strategy, set them to the max X/Y that can be travelled from the center of the bed (assumed to be 0,0).